Add working twitter, facebook, link box, and AddThis buttons.
Facebook descriptions and thumbnails can be made better after we deploy,
because Facebook scrapes the shared URL for metadata.

AddThis doesn’t work so well -- e.g. it adds @addthis to tweets instead of
@datacollective -- so we turn those services off in AddThis.

fb-fw.png is the original Fireworks graphic with vector data and all that.
The -32 and -256 versions are optimized.